@@ -5,62 +5,64 @@
   "scripts": {
     "ng": "nx",
     "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
-    "start": "nx serve",
+    "start": "nx run flight-app:serve-with-api",
     "build": "nx build",
     "test": "nx test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~14.0.0",
-    "@angular/common": "~14.0.0",
-    "@angular/compiler": "~14.0.0",
-    "@angular/core": "~14.0.0",
-    "@angular/forms": "~14.0.0",
-    "@angular/platform-browser": "~14.0.0",
-    "@angular/platform-browser-dynamic": "~14.0.0",
-    "@angular/router": "~14.0.0",
-    "@nestjs/common": "^8.0.0",
-    "@nestjs/core": "^8.0.0",
-    "@nestjs/platform-express": "^8.0.0",
-    "@nrwl/angular": "14.4.2",
-    "reflect-metadata": "^0.1.13",
-    "rxjs": "~7.4.0",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "@angular/animations": "14.1.0",
+    "@angular/common": "14.1.0",
+    "@angular/compiler": "14.1.0",
+    "@angular/core": "14.1.0",
+    "@angular/forms": "14.1.0",
+    "@angular/platform-browser": "14.1.0",
+    "@angular/platform-browser-dynamic": "14.1.0",
+    "@angular/router": "14.1.0",
+    "@nestjs/common": "8.0.0",
+    "@nestjs/core": "8.0.0",
+    "@nestjs/platform-express": "8.0.0",
+    "@ng-web-apis/common": "^2.0.1",
+    "@nrwl/angular": "14.4.3",
+    "reflect-metadata": "0.1.13",
+    "rxjs": "7.4.0",
+    "tslib": "2.0.0",
+    "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~14.0.0",
-    "@angular-eslint/eslint-plugin": "~14.0.0",
-    "@angular-eslint/eslint-plugin-template": "~14.0.0",
-    "@angular-eslint/template-parser": "~14.0.0",
-    "@angular/cli": "~14.0.0",
-    "@angular/compiler-cli": "~14.0.0",
-    "@angular/language-service": "~14.0.0",
-    "@nestjs/schematics": "^8.0.0",
-    "@nestjs/testing": "^8.0.0",
-    "@nrwl/cli": "14.4.2",
-    "@nrwl/cypress": "14.4.2",
-    "@nrwl/eslint-plugin-nx": "14.4.2",
-    "@nrwl/jest": "14.4.2",
-    "@nrwl/linter": "14.4.2",
-    "@nrwl/nest": "14.4.2",
-    "@nrwl/node": "14.4.2",
-    "@nrwl/nx-cloud": "latest",
-    "@nrwl/workspace": "14.4.2",
+    "@angular-devkit/build-angular": "14.0.0",
+    "@angular-eslint/eslint-plugin": "14.0.0",
+    "@angular-eslint/eslint-plugin-template": "14.0.0",
+    "@angular-eslint/template-parser": "14.0.0",
+    "@angular/cli": "14.0.0",
+    "@angular/compiler-cli": "14.1.0",
+    "@angular/language-service": "14.1.0",
+    "@nestjs/schematics": "8.0.0",
+    "@nestjs/testing": "8.0.0",
+    "@nrwl/cli": "14.4.3",
+    "@nrwl/cypress": "14.4.3",
+    "@nrwl/eslint-plugin-nx": "14.4.3",
+    "@nrwl/jest": "14.4.3",
+    "@nrwl/linter": "14.4.3",
+    "@nrwl/nest": "14.4.3",
+    "@nrwl/node": "14.4.3",
+    "@nrwl/nx-cloud": "14.2.0",
+    "@nrwl/workspace": "14.4.3",
     "@types/jest": "27.4.1",
     "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "^5.29.0",
-    "@typescript-eslint/parser": "^5.29.0",
-    "cypress": "^9.1.0",
-    "eslint": "~8.15.0",
+    "@typescript-eslint/eslint-plugin": "5.29.0",
+    "@typescript-eslint/parser": "5.29.0",
+    "cypress": "9.1.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-cypress": "^2.10.3",
+    "eslint-plugin-cypress": "2.10.3",
     "jest": "27.5.1",
-    "jest-preset-angular": "~11.1.2",
-    "nx": "14.4.2",
-    "prettier": "^2.6.2",
+    "jest-preset-angular": "11.1.2",
+    "nx": "14.4.3",
+    "prettier": "2.6.2",
     "ts-jest": "27.1.4",
-    "ts-node": "~10.8.0",
-    "typescript": "~4.7.2"
+    "ts-node": "10.8.0",
+    "type-fest": "^2.17.0",
+    "typescript": "4.7.2"
   }
 }
