@@ -3,14 +3,27 @@ import { SidebarComponent } from './sidebar/sidebar.component';
 import { HttpClientModule } from '@angular/common/http';
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
 
 import { AppComponent } from './app.component';
+import { AppRoutingModule } from './app-routing.module';
+import { HomeComponent } from './home/home.component';
+import { FlightsModule } from './flights/flights.module';
+import { FlightSearchModule } from './flight-search/flight-search.module';
 
 @NgModule({
-  imports: [BrowserModule, HttpClientModule, FormsModule],
-  declarations: [AppComponent, SidebarComponent, NavbarComponent],
-  providers: [],
+  imports: [
+    BrowserModule,
+    HttpClientModule,
+    AppRoutingModule,
+    FlightsModule,
+    FlightSearchModule,
+  ],
+  declarations: [
+    AppComponent,
+    SidebarComponent,
+    NavbarComponent,
+    HomeComponent,
+  ],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
